+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- cli: add csv export option to tables query command
+
 ### Changes
 - cli: table uri lazy validation in app configuration
 - web: table uri lazy validation in app configuration